Preliminary support for Tornado 5+
This involved 3 main changes:
1. restructuring how the ioloop is handled for Tornado 4.1+ standards.
Previously, ioloop instances were manually managed, but this was
deprecated in 4.1 and removed somewhere around 5. What I am not sure of
is whether the previous code, when run on Tornado 3ish, entailed many
ioloops in one process (something not actually supported by Tornado).

2. Rewriting GameDataHandler so that it doesn't override `get`.
Apparently this is a challenging thing to do right; the new version
parses the version number out of the url on request.

3. Restructuring shutdown/stop for better signal handling. The old code
looks slightly broken to me, and I suspect that issues here may be the
source of some issues where cleanup doesn't happen properly.

There are also a bunch of other miscellaneous fixes, some pulled from
webtiles-changes. In my testing, this actually seems to all be backwards
compatible to Tornado 3.2.1?? Which is a bit too good to be true...