Formicids cannot cast Swiftness
This has been a long-standing bone of contention in vanilla (see Mantis
issue 7786) because the original formicid restrictions were a somewhat
arbitary duplication of the effects of the now-gone amulet of stasis.

If formicids can't benefit from Haste, well Swiftness is Haste Lite.

This change is borrowed from Stoat Soup.