Run timer callbacks in main thread
Previously, they were run in a separate thread spawned by SDL, which
makes interacting with other code difficult, since most other code
assumes single-threaded access.

Note that this change queues an SDL event, but only the input loop in
ui.cc actually handles this event by calling the callback. This is fine,
because currently the only use of this API is to show tooltips after
a delay, and the callback function for that worked by queueing an SDL
event, so the callback now does nothing and the actual function is
performed as a side effect.

This may seem like a step backwards for code clarity, but it's a
necessary step to removing the input loop in tilesdl.

This was previously reverted at 431bb856fd. I'm adding it now after
manual testing with both a crosscompile in Wine and in a Windows VM
with MSYS2, both with mingw-w64-1686. Neither showed any lag or crashes.