Correct and refactor monster enchant durations
The monster enchants Aura of Brilliance (used by deep elf high priests),
Empowered Spells (conferred by Aura of Brilliance), and Frozen (debuff
from Flash Freeze) monster enchants have have default durations that
incorrectly use AUT scale for a count of turns. This meant that Aura of
Brilliance and Frozen enchants were added with 10 times the intended
number of turns. This doesn't have much gameplay significance in
practice, since deep elf high priests can just recast the buff when it
expires and Frozen is generally not applied to monsters. The default
duration for Empowered Spells was unused, as this duration was applied
elsewhere in a way that was almost correct. None of these enchants
factor in the monster's intrinsic speed like they are supposed to.

This commit corrects the default durations to the right scale and
incorporates monster intrinsic speed. The duration application of
Empowered Spells now uses the default duration instead of having an
independent calculation.