Extend alias functionality for database lookups
Currently, descriptions for abilities, spells, and dungeon features can
be reused via aliases, d52d0508. But it's not possible to use an alias
and then add details to it.

This commit adds support for a new type of marker for database entries,
[[foo]]. Any [[foo]] marker in a database entry will be replaced with a
corresponding entry's text before being displayed to the player.