Fix unrand uniqueness for acquirement (AliasTheSpectator)
Scrolls of acquirement copy item definitions for the generated items and
remove the created items, only re-making the selected item when the
player makes a choice. For unrands, the item def has to have its
uniqueness status cleared, but this wasn't properly reinstated if the
unrand ended up being chosen by the player. Hence it was possible to
generate a specific unrand again after acquiring it, either through
another acquirement or as random loot.

With this commit, we properly set the unrand's uniqueness status when
making the final item in acquirement, making it actually unique.