Modify five vaults in large.des to add tactics
The current arrival vault standard is for arrival vaults to have
multiple exits, a hatch, or a large enough internal space for relevant
tactics. I went through large.des picking out clear offenders (using the
simplified rubric that anywhere a single gnoll could spawn should still
leave you with more options than 'tab and hope you win.')

elethiomel_arrival_infinity_welcomes_careful_drivers: This vault has an
issue if a gnoll spawns in the closet just outside the starting room. I
marked the contents of the infinity-symbol as no_monster_gen.

erik_arrival_gehennom: This vault explictly places a goblin and kobold
(with standard weapon generation) right outside the starting room. Many
starts are incapable of fighting a goblin and kobold with a bad weapon
spawn in this layout. This vault is also spoilery and doesn't even match
any vaults in Nethack (which it's supposed to be an homage to.) I
removed it.

kilobyte_arrival_exit_the_forest: This vault is effectively a single
long corridor with monster-gen on; if a gnoll spawns and you can't fight
him, you're screwed. I added a path through the woods in the northwest
so you have a large pillar to work with, and a second exit past the
water to the south-east.

arrival_zaba_pyramid: Each of the possible starting rooms are
single-exit, and monster-gen is on. So it's possible for a gnoll to
spawn right in front of the door to your room, and completely block you
in. I marked the small intermediary rooms as no_monster_gen.

minmay_arrival_shoal_huts: This vault is another case with single-exit
entry rooms which can be blocked by an unlucky gnoll spawn. I added a
hatch to each entry room as an additional option.

[ Committer's note: wrapped message to 80 cols. Closes #1290 ]