Add a pointer check to actor::collide
This is to address a crash triggered by primal wave. It is in principle
a bug that the agent would be unset when a collision is triggered by
primal wave via phial, but I can't actually replicate the exact
circumstances that cause it to be unset. I think it may be able to
happen, somehow, when the player is confused?

In any case, this code can't assume that a pointer argument will not be
null like this.