Disallow evoking lamp/phial while confused
These two are almost the last remaining holdover of evocables that could
be used while confused. (I notice that tremorstones also work.) These
silently applied target fuzzing (see 657136525ae) in this case; if these
were to be usable while confused in modern crawl, I think they would
need to indicate to the player somehow what might happen. I ran across
this in a crash where an extremely desparate player was constricted and
chain-confused by golden eyes tried to target the constricting naga with
the phial, and fuzzing cause the wave to go the other direction and hit
a friendly demon. Luckily for this player, the game crashed during
knockback on the demon.

See cccf84277269a for wands, and 6178f7666a6f for rods. Unlike these
cases, I do think it's possible that lamp/phial might not be completely
useless with fuzzed targeting, so I wouldn't be opposed to bringing this
back if someone can solve the UI problem.