Log an error on `idle_connection_timeout` for old Tornado versions
Previously, this just prevented the server from starting and users of
old Tornado versions were instructed to install a patched version of
Tornado. This takes an alternative strategy of just ignoring the timeout
parameter, which is implemented on Tornado 4 and later. Hopefully this
won't cause problems for any active servers...