Tornado 2.4 compatibility hacks
It seems that many or all servers may be running the patched version of
Tornado 2.4 found in edlothiel's git repository. To ease transition to
modern versions, this commit implements several compatibility patches
needed to not completely break these servers. Everything in this commit
is ugly and not supposed to be permanent. (We know how that always
goes...)