Add a rudimentary serverwide announce facility to webtiles
This lets DGL admin users make server-wide announcements, e.g. if the
server is going to shut down. The announcement appears both in game logs
and in the chat window. The facility is pretty basic, but functional,
and is designed with the intent that it would be *extremely* rarely
used.

Some questions:

* should the admin panel remain part of client.html?
* Is DGL admin the right permissions level?
* what else should be added to the new admin panel?
* is announcing to both the game and chat too much?
* will this work with ssh games? (I think so?)