Refactor init_user to be asynchronous
This code calls a script via a subprocess. On CAO, this script takes
about 600ms to run on average for new users, and so was triggering
timeouts on nearly every new user (potentially on logins as well),
presumably blocking the server as well. This refactors things so that
the exit condition is checked and handled in a callback rather than in
the same function after the subprocess returns. (This code be cleaner in
pure python 3, but I couldn't get any of those solutions to fully work
with the needed backwards compatibility.)