Some tweaks to reverse_door_vault
This variant of door_vault is unique in that it places a single upstairs
in an enclosed area, with the doors revealing this area fully. I've
modified it to use the common door vault setup function, so now it
places appropriate loot. I don't change the levels of placement, since
this one isn't interchangeable with the other variants.

The monster set and count are fine, since it only uses 8 and 9 spawns in
the same numbers as door_vault. Due to presence of a staircase, this
vault in its current form will always be less prone to killing people.
Hence I'll probably revamp it to add traps and/or specialized monster
placement.