Fix a piercing + omnireflection crash
If a piercing ranged attack is reflected by warlocks mirror, kills the
thrower, and then hits a monster after that, the game would crash,
because the attacker is invalid. This fix lets the piercing ammo
continue without crashing by using YOU_FAULTLESS as the attacker in
this case.

It isn't quite right, since the to-hit is now calculated based on a fake
YOU_FAULTLESS rather than a monster that the beam/ranged_attack objects
no longer has access to. I can see a couple of possibilities; we could
just make a copy of the monster (extremely heavy), we could somehow only
clean up the monster with a fineff (tricky to implement), or rewrite
this code to defer dead monster cleanup until the end of monster
handling (also tricky to implement, but probably how I would do this if
writing all of this code from scratch -- this would have avoided many,
many bugs over the years). I can't see any simpler options than these
right now.

Another change might be to have the ammo vanish when the monster dies,
on the premise that piercing is a magic effect that requires the thrower
to be focusing.