Fix lint errors in util.py, and add type hints
This is made somewhat more complicated by Tornado hinting some functions
as accepting a str with a default of None; necessitating several type
ignore comments. Some more were added to quickly deal with the forced
addition of an attribute to instances of one of Tornado's classes (not
a great idea...).

This also removes a race condition from the FileTailer class.