Suppress builder dprf spam from main message window by default
These logs are saved with the level via a message_tee and are accessible
via &ctrl-l in wizmode. This is basically because incremental pregen in
a fulldebug build produces a huge amount of log spam whenever any
non-trivial level change happens.  Hopefully this doesn't step on some
workflow I'm not thinking of, but it's configurable (and could be made
more configurable).