Remove miscasts from Xom effects
Now that miscasts effects are focused and meant to punish spell misuse
they are less useful as "random wacky junk" in the Xom effect roster.
Removing the lethality margin also significantly simplifies things.

Xom effects probably need a more thorough looking-at, but that is beyond
the scope of this change.

BcadrenCrawl: Though I may add them back, the restructuring means they
wouldn't work as is anyways so let them be stripped for now.