Shut down cleanly on SIGTERM
Treat SIGTERM the same as SIGHUP -- eg gracefully save and quit.

When a SSH client disconnects from a remote session (eg, with
dgamelaunch), the session processes receive SIGHUP. But if a local user
runs eg `killall crawl`, the processes are sent SIGTERM which crashes
the game. Most job control systems (eg systemd, docker) send SIGTERM, so
it makes sense to handle this with the logic that already exists for
SIGHUP.

Don't rename any of the functions or variables relating to this, eg
handle_hangup or seen_hups. That can be done in a future commit, if
desired.