Remove code for old potion types
This commit adds an CASE_REMOVED_POTIONS macro, intended to be used in
switch statements, that will die() with an error for any removed potion
type. It then improves the tags.cc conversion of old potions; now, *all*
removed potions are converted, with a runtime crash on load guarding
against forgetting to update this.

This means that no saves can have these potions in the game. We follow
up by removing reams of remaining cruft that's no longer necessary as a
result.