Fix Trog weapon gifting (elliptic)
Trog weapon gifting was partially broken by 6d7e32a81, which
accidentally removed the two successful random rolls needed if piety was
at the 5* threshold required. This lead to excessive Trog gifts as well
as it being impossible to actually reach 6*, since the gift timeout
would consume most piety. This bug did not affect Okawaru weapon gifts,
since the rolls were correctly applied.

This code block was structured in a confusing way to begin with, so I've
both fixed and refactored it to make the logic clearer.