Fix targeters sometimes being too short for Barachi (#11833)
in_bounds() takes an absolute pair of coordinates, but the view code
was passing in player-relative coordinates. This caused ray cells to
be hidden along the top and left edges, where ep.y and ep.x were zero
respectively, and thus fell out of the range of valid map cells when
interpreted as a grid location.