Generalize b977e01cb5bc to other penetration crashes
This guarantees that a ranged attack will be attempted for penetration
weapons regardless of the state of the agent. In b977e01cb5bc I used
YOU_FAULTLESS as the agent for cases where reflection kills the agent,
and this commit generalizes this to use YOU_FAULTLESS for all cases
where the agent is gone. This is conceptually odd, but better than
crashing; and I think it is more consistent behavior than alternatives
like causing the penetration attack to stop at that point, or letting it
continue without damaging anything.